+++ src/main/java/org/dynamide/restreplay/mutators/VarMutator.java
@@ -75,7 +75,6 @@ public class VarMutator extends ContentMutator implements IMutator {
         return safeID();
     }
 
-    @Override
     public int getIndex(){
         return index-1;
     }